refactor(hero): create seperate component, split into smaller files
