+++ components/hero-slider/slides-data.tsx
@@ -1,0 +1,91 @@
+export const slidesData = [
+  {
+    title: {
+      text: `Introducing Zeitgeist Avatar`,
+      size: `3.5rem`,
+      color: `#FFF`,
+    },
+    bg: "/carousel/slide-1.jpeg",
+    cta1: "Protocol Exclusive",
+    link1: "https://blog.zeitgeist.pm/",
+    color1: {
+      primary: "#E90303",
+      secondary: "#FFF",
+      border: "#E90303",
+    },
+    cta2: "Read Post",
+    link2: "https://blog.zeitgeist.pm/zeitgeist-newsletter-002/",
+    color2: {
+      primary: "transparent",
+      secondary: "#FFF",
+      border: "#FFF",
+    },
+  },
+  {
+    title: {
+      text: `Exiled Racers`,
+      size: `4rem`,
+      color: `#FFF`,
+    },
+    bg: "/carousel/slide-2.jpeg",
+    cta1: "New Markets",
+    link1: "https://shapethefuture.zeitgeist.pm",
+    color1: {
+      primary: "#000",
+      secondary: "#E90303",
+      border: "#E90303",
+    },
+    cta2: "",
+    link2: "https://shapethefuture.zeitgeist.pm",
+    color2: {
+      primary: "",
+      secondary: "",
+      border: "",
+    },
+  },
+  {
+    title: {
+      text: `ZTG is now traded on DEXs`,
+      size: `4rem`,
+      color: `#FFF`,
+    },
+    bg: "/carousel/slide-3.jpeg",
+    cta1: "Visit In-App Exchange",
+    link1: "https://shapethefuture.zeitgeist.pm",
+    color1: {
+      primary: "#F7FF58",
+      secondary: "#000",
+      border: "#F7FF58",
+    },
+    cta2: "",
+    link2: "https://shapethefuture.zeitgeist.pm",
+    color2: {
+      primary: "",
+      secondary: "",
+      border: "",
+    },
+  },
+  {
+    title: {
+      text: `Presenting Zeitgeist’s 
+      "Rikiddo Scoring Rule"`,
+      size: `3rem`,
+      color: `#FFF`,
+    },
+    bg: "/carousel/slide-4.jpeg",
+    cta1: "New Markets",
+    link1: "https://shapethefuture.zeitgeist.pm",
+    color1: {
+      primary: "#F7FF58",
+      secondary: "#000",
+      border: "#F7FF58",
+    },
+    cta2: "",
+    link2: "https://shapethefuture.zeitgeist.pm",
+    color2: {
+      primary: "",
+      secondary: "",
+      border: "",
+    },
+  },
+];
